+++ tailwind.config.typography.js
@@ -15,28 +15,24 @@ module.exports = {
             typography: (theme) => ({
                 DEFAULT: {
                     css: {
-                        "--tw-prose-body": theme("colors.neutral.DEFAULT"),
-                        "--tw-prose-headings": theme("colors.neutral.DEFAULT"),
-                        "--tw-prose-lead": theme("colors.neutral.DEFAULT"),
-                        "--tw-prose-links": theme("colors.neutral.DEFAULT"),
-                        "--tw-prose-bold": theme("colors.neutral.DEFAULT"),
-                        "--tw-prose-counters": theme("colors.neutral.DEFAULT"),
-                        "--tw-prose-bullets": theme("colors.neutral.DEFAULT"),
-                        "--tw-prose-hr": theme("colors.neutral.DEFAULT"),
-                        "--tw-prose-quotes": theme("colors.neutral.DEFAULT"),
+                        "--tw-prose-body": theme("colors.white"),
+                        "--tw-prose-headings": theme("colors.white"),
+                        "--tw-prose-lead": theme("colors.white"),
+                        "--tw-prose-links": theme("colors.white"),
+                        "--tw-prose-bold": theme("colors.white"),
+                        "--tw-prose-counters": theme("colors.white"),
+                        "--tw-prose-bullets": theme("colors.white"),
+                        "--tw-prose-hr": theme("colors.white"),
+                        "--tw-prose-quotes": theme("colors.white"),
                         "--tw-prose-quote-borders": theme(
                             "colors.primary.DEFAULT / 1"
                         ),
-                        "--tw-prose-captions": theme("colors.neutral.DEFAULT"),
-                        "--tw-prose-code": theme("colors.neutral.DEFAULT"),
+                        "--tw-prose-captions": theme("colors.white"),
+                        "--tw-prose-code": theme("colors.white"),
                         "--tw-prose-pre-code": theme("colors.white"),
-                        "--tw-prose-pre-bg": theme("colors.neutral.DEFAULT"),
-                        "--tw-prose-th-borders": theme(
-                            "colors.neutral.DEFAULT"
-                        ),
-                        "--tw-prose-td-borders": theme(
-                            "colors.neutral.DEFAULT"
-                        ),
+                        "--tw-prose-pre-bg": theme("colors.white"),
+                        "--tw-prose-th-borders": theme("colors.white"),
+                        "--tw-prose-td-borders": theme("colors.white"),
                         "--tw-prose-invert-body": theme("colors.white"),
                         "--tw-prose-invert-headings": theme("colors.white"),
                         "--tw-prose-invert-lead": theme("colors.white"),
@@ -50,9 +46,7 @@ module.exports = {
                             theme("colors.white"),
                         "--tw-prose-invert-captions": theme("colors.white"),
                         "--tw-prose-invert-code": theme("colors.white"),
-                        "--tw-prose-invert-pre-code": theme(
-                            "colors.neutral.DEFAULT"
-                        ),
+                        "--tw-prose-invert-pre-code": theme("colors.white"),
                         "--tw-prose-invert-pre-bg": theme("colors.white"),
                         "--tw-prose-invert-th-borders": theme("colors.white"),
                         "--tw-prose-invert-td-borders": theme("colors.white"),